fix(styled-components): globalStyle
