+++ src/components/ThemeProvider/index.js
@@ -1,12 +1,12 @@
 import React, { useContext, useState } from 'react';
 import { createGlobalStyle, ThemeProvider } from 'styled-components';
+import 'modern-css-reset';
 
 import Toggle from '../Toggle';
 import defaultTheme, { dark, light } from '../../themes';
 import tokens from '../../tokens';
 
-import './reset';
-import Fonts from './fonts';
+import useGoogleFont from './useGoogleFont';
 
 const GlobalStyle = createGlobalStyle`  
 	html {
@@ -40,6 +40,9 @@ const GlobalStyle = createGlobalStyle`
 const ThemeContext = React.createContext({});
 
 const TalendThemeProvider = ({ theme = defaultTheme, children }) => {
+	useGoogleFont(
+		'https://fonts.googleapis.com/css2?family=Open+Sans:ital,wght@0,300;0,400;0,600;0,700;0,800;1,300;1,400;1,600;1,700;1,800&family=Inconsolata:wght@300;400;500;600;700;800;900&display=swap',
+	);
 	const [selectedTheme, setSelectedTheme] = useState(theme);
 	React.useEffect(() => {
 		setSelectedTheme(theme);
@@ -47,7 +50,6 @@ const TalendThemeProvider = ({ theme = defaultTheme, children }) => {
 	const switchTheme = newTheme => setSelectedTheme(newTheme);
 	return (
 		<ThemeContext.Provider value={{ switchTheme, theme: selectedTheme }}>
-			<Fonts />
 			<ThemeProvider theme={selectedTheme}>{children}</ThemeProvider>
 		</ThemeContext.Provider>
 	);
