--- src/components/ThemeProvider/reset.ts
@@ -1,1 +1,0 @@
-import 'modern-css-reset';
