+++ tsconfig.json
@@ -1,11 +1,18 @@
 {
   "extends": "./node_modules/@talend/scripts-config-typescript/tsconfig.json",
-  "include": ["src/*", "custom.d.ts"],
-  "exclude": ["dist", "node_modules"],
+  "include": [
+    "src/*",
+    "custom.d.ts"
+  ],
+  "exclude": [
+    "dist",
+    "node_modules"
+  ],
   "compilerOptions": {
     "declaration": true,
-    "jsx": "react",
+    "jsx": "preserve",
     "target": "ES5",
-    "module": "CommonJs"
+    "module": "CommonJs",
+    "noEmit": true
   }
 }
